# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Produtos
  version: "1.0.0"
  description: >
    API RESTful para gerenciamento de produtos.  
    Implementa autenticação JWT e segue boas práticas REST.
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor local (desenvolvimento)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        msg:
          type: string
          example: Mensagem de erro
paths:
  /auth/login:
    post:
      summary: Autentica e gera token JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Token emitido com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        '422':
          description: Campos ausentes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...